git_rebase_branches: ask the opposite question
Instead of asking "wanna skip?", now ask "wanna rebase?".
This way the user just needs to hit enter in order to do the rebase.
And 'enter' is commonly known to mean 'yes'.

A branch is *not* rebased only if the user types an explicit 'n'
(or 'N', or 'no', or 'NO').